Fix: didn't check that a task is archived before deleting it

+++ src/services/task-service.js
@@ -43,7 +43,7 @@ class TaskService {
 
   delete(taskId) {
     const task = this.findByIdOrFail(taskId);
-    if (task) {
+    if (task && task.isArchived) {
       this.taskRepository.delete(task)
       this.analyticService.update(task.id)
     } else {
